Split persistence context tests from persistence tests (#844).
