+++ legend-engine-xt-persistence-grammar/src/test/java/org/finos/legend/engine/language/pure/dsl/persistence/grammar/test/TestPersistenceGrammarParser.java
@@ -61,116 +61,6 @@ public class TestPersistenceGrammarParser extends TestGrammarParser.TestGrammarP
     }
 
     /**********
-     * persistence context
-     **********/
-
-    @Test
-    public void persistenceContextPersistence()
-    {
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "}\n", "PARSER error at [3:1-5:1]: Field 'persistence' is required");
-
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "}\n", "PARSER error at [3:1-7:1]: Field 'persistence' should be specified only once");
-    }
-
-    @Test
-    public void persistenceContextPersistencePlatform()
-    {
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "}\n");
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "  platform: Default;\n" +
-                "}\n");
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "  platform: Unknown;\n" +
-                "}\n", "PARSER error at [6:13-19]: Unsupported persistence platform type 'Unknown'");
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "  platform: Default;\n" +
-                "  platform: Default;\n" +
-                "}\n", "PARSER error at [3:1-8:1]: Field 'platform' should be specified only once");
-    }
-
-    @Test
-    public void persistenceContextServiceParameters()
-    {
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "}\n");
-
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "  serviceParameters: [foo='hello',bar=1];\n" +
-                "  serviceParameters: [foo='hello',bar=1];\n" +
-                "}\n", "PARSER error at [3:1-8:1]: Field 'serviceParameters' should be specified only once");
-    }
-
-    @Test
-    public void persistenceContextSinkConnection()
-    {
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "}\n");
-
-        test("###Persistence\n" +
-                "\n" +
-                "PersistenceContext test::TestPersistenceContext \n" +
-                "{\n" +
-                "  persistence: test::TestPersistence;\n" +
-                "  serviceParameters: [foo='hello',bar=1];\n" +
-                "  sinkConnection: #{\n" +
-                "    RelationalDatabaseConnection {\n" +
-                "      store: test::TestDatabase;\n" +
-                "      type: H2;\n" +
-                "      specification: LocalH2{};\n" +
-                "      auth: Test;\n" +
-                "    }\n" +
-                "  }#;\n" +
-                "  sinkConnection: #{\n" +
-                "    RelationalDatabaseConnection {\n" +
-                "      store: test::TestDatabase;\n" +
-                "      type: H2;\n" +
-                "      specification: LocalH2{};\n" +
-                "      auth: Test;\n" +
-                "    }\n" +
-                "  }#;\n" +
-                "}\n", "PARSER error at [3:1-23:1]: Field 'sinkConnection' should be specified only once");
-    }
-
-    /**********
      * persistence
      **********/
